feat(config,ui): add scroll_past_list option, fix issue with last post stats not visible
Before, the stats (points, comments, author & date) of the last post in
the list were not visible. They were there, except not rendered by the
way list scrolling was implemented. The scroll_past_list options is
added to fix this. It needed to be optional as it's not very visibly
pleasing when enabled. As functionality is more important than being
visibly pleasing in my opinion, this option is enabled by default.